[#59] refactor : 문제 생성 부분 필수 주석, 코드 말고 제거
