+++ src/main/java/org/unilab/improfessorbe/domain/problem/service/ProblemService.java
@@ -56,7 +56,6 @@ public class ProblemService {
 
 			log.info("문제 생성 및 캐시 저장 완료: 총 {}개 문제, 다운로드 키: {}", responses.size(), downloadKey);
 
-			//3. 유저 무료 횟수 차감
 			userService.decrementFreeCount(userId);
 
 			return ProblemGenerationResponse.of(downloadKey, responses);
@@ -82,7 +81,6 @@ public class ProblemService {
 		// 3. 파일명 생성
 		String fileName = createDownloadFileName();
 
-		// 4. 로깅
 		log.info("문제 PDF 생성 완료: key={}, 파일명={}, 문제수={}",
 			downloadKey, fileName, cachedData.getProblems().size());
 
@@ -103,33 +101,6 @@ public class ProblemService {
 		}
 	}
 
-	/*public List<ProblemResponse> createProblem(List<MultipartFile> conceptFiles, List<MultipartFile> formatFiles) {
-		try {
-			String conceptContent = fileParseService.parseFileList(conceptFiles, "개념");
-			String formatContent = "";
-			if (formatFiles != null) {
-				formatContent = fileParseService.parseFileList(formatFiles, "형식");
-			}
-
-			log.info("개념 파일 글자수: {}개 / 형식 파일 글자수: {}개", conceptContent.length(), formatContent.length());
-
-			String problemText = geminiApiClient.generateProblems(conceptContent, formatContent);
-			List<Problem> problems = problemTextParser.parseProblemText(problemText);
-
-			List<ProblemResponse> responses = new ArrayList<>();
-			for (Problem problem : problems) {
-				responses.add(ProblemResponse.toResponse(problem));
-			}
-
-			return responses;
-
-		} catch (CustomException e) {
-			throw e;
-		} catch (Exception e) {
-			throw new CustomException(ErrorCode.PROBLEM_CREATION_FAILED);
-		}
-	}*/
-
 	public List<ProblemResponse> createProblemWithMl(List<MultipartFile> conceptFiles,
 		List<MultipartFile> formatFiles) {
 		try {
@@ -142,7 +113,6 @@ public class ProblemService {
 			log.info("개념 파일 글자수: {}개 / 형식 파일 글자수: {}개",
 				conceptContent.length(), formatContent.length());
 
-			//ML 추가
 			ConceptExtractionResult result = conceptExtractorService.extractConcepts(conceptContent);
 			String conceptExtraction = result.toFormattedString();
